# docs/intel接口文档-中文版本-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Host Service",
        "description": "主机管理和WebSocket实时通信服务",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/host/admin/ota/list": {
            "get": {
                "tags": [
                    "管理后台-OTA管理"
                ],
                "summary": "查询 OTA 配置列表",
                "description": "查询 sys_conf 表中 conf_key = 'ota', state_flag = 0, del_flag = 0 的全部数据",
                "operationId": "list_ota_configs_api_v1_host_admin_ota_list_get",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_AdminOtaListResponse_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/admin/ota/deploy": {
            "post": {
                "tags": [
                    "管理后台-OTA管理"
                ],
                "summary": "下发 OTA 配置",
                "description": "下发 OTA 配置到所有连接的 Host，更新 sys_conf 表并广播消息",
                "operationId": "deploy_ota_config_api_v1_host_admin_ota_deploy_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminOtaDeployRequest",
                                "description": "OTA 下发请求数据"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "下发成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_AdminOtaDeployResponse_"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OTA 配置不存在"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminOtaConfigInfo": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "配置ID（主键）"
                    },
                    "conf_ver": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Ver",
                        "description": "配置版本号"
                    },
                    "conf_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Name",
                        "description": "配置名称"
                    },
                    "conf_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Url",
                        "description": "OTA 包下载地址"
                    },
                    "conf_md5": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Md5",
                        "description": "OTA 包 MD5 校验值"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "AdminOtaConfigInfo",
                "description": "管理后台 OTA 配置信息"
            },
            "AdminOtaDeployRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "title": "Id",
                        "description": "配置ID（主键）"
                    },
                    "conf_ver": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Conf Ver",
                        "description": "配置版本号"
                    },
                    "conf_name": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Conf Name",
                        "description": "配置名称"
                    },
                    "conf_url": {
                        "type": "string",
                        "title": "Conf Url",
                        "description": "OTA 包下载地址（字符串，允许任意格式）"
                    },
                    "conf_md5": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 32,
                                "minLength": 32,
                                "pattern": "^[a-fA-F0-9]{32}$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Md5",
                        "description": "OTA 包 MD5 校验值（32位十六进制，可选）"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "conf_ver",
                    "conf_name",
                    "conf_url"
                ],
                "title": "AdminOtaDeployRequest",
                "description": "管理后台 OTA 下发请求模式"
            },
            "AdminOtaDeployResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "配置ID（主键）"
                    },
                    "conf_ver": {
                        "type": "string",
                        "title": "Conf Ver",
                        "description": "配置版本号"
                    },
                    "conf_name": {
                        "type": "string",
                        "title": "Conf Name",
                        "description": "配置名称"
                    },
                    "conf_url": {
                        "type": "string",
                        "title": "Conf Url",
                        "description": "OTA 包下载地址"
                    },
                    "conf_md5": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conf Md5",
                        "description": "OTA 包 MD5 校验值（可选）"
                    },
                    "broadcast_count": {
                        "type": "integer",
                        "title": "Broadcast Count",
                        "description": "广播消息成功发送的主机数量"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "conf_ver",
                    "conf_name",
                    "conf_url",
                    "broadcast_count"
                ],
                "title": "AdminOtaDeployResponse",
                "description": "管理后台 OTA 下发响应模式"
            },
            "AdminOtaListResponse": {
                "properties": {
                    "ota_configs": {
                        "items": {
                            "$ref": "#/components/schemas/AdminOtaConfigInfo"
                        },
                        "type": "array",
                        "title": "Ota Configs",
                        "description": "OTA 配置列表"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total",
                        "description": "配置总数"
                    }
                },
                "type": "object",
                "required": [
                    "ota_configs",
                    "total"
                ],
                "title": "AdminOtaListResponse",
                "description": "管理后台 OTA 配置列表响应模式"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Result_AdminOtaDeployResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdminOtaDeployResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[AdminOtaDeployResponse]"
            },
            "Result_AdminOtaListResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdminOtaListResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[AdminOtaListResponse]"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}